Split some test classes in multiple files
